#95 Move specifications of Dataset to specs module
