+++ ndradex/__init__.py
@@ -1,4 +1,12 @@
-__all__ = ["LAMDA", "lamda", "get_lamda", "nd", "radex", "run"]
+__all__ = [
+    "LAMDA",
+    "lamda",
+    "get_lamda",
+    "nd",
+    "radex",
+    "run",
+    "specs",
+]
 __version__ = "0.3.1"
 
 
@@ -6,6 +14,7 @@ __version__ = "0.3.1"
 from . import lamda
 from . import nd
 from . import radex
+from . import specs
 from .lamda import LAMDA, get_lamda
 from .nd import run
