+++ ndradex/nd.py
@@ -3,43 +3,25 @@ __all__ = ["run"]
 
 # standard library
 from csv import writer as csv_writer
-from dataclasses import dataclass, field
 from itertools import product
 from os import PathLike
 from tempfile import NamedTemporaryFile, TemporaryFile
-from typing import Any, Collection, IO, Iterator, Literal, TypeVar
+from typing import Any, Collection, IO, Iterator, TypeVar
 
 
 # dependencies
-import numpy as np
 import pandas as pd
 import xarray as xr
-from astropy.units import Quantity
 from tqdm import tqdm
-from xarray_dataclasses import AsDataset, DataModel, Attr, Coordof, Data, Dataof
 from .lamda import get_lamda
 from .radex import RadexInput, runmap, to_input
+from .specs import EmptySet
 
 
 # type hints
 T = TypeVar("T")
 Multiple = Collection[T] | T
 StrPath = PathLike[str] | str
-VarDims = tuple[
-    Literal["transition"],
-    Literal["T_kin"],
-    Literal["n_H2"],
-    Literal["n_pH2"],
-    Literal["n_oH2"],
-    Literal["n_e"],
-    Literal["n_H"],
-    Literal["n_He"],
-    Literal["n_Hp"],
-    Literal["T_bg"],
-    Literal["N"],
-    Literal["dv"],
-    Literal["radex"],
-]
 
 
 # constants
@@ -206,226 +188,3 @@ def walk_dims(dataset: xr.Dataset) -> Iterator[dict[str, Any]]:
 
     for values in product(*dims.values()):
         yield dict(zip(dims.keys(), values))
-
-
-class Units:
-    """Convert data with units to given units."""
-
-    data: Any
-    units: Any
-
-    def __post_init__(self) -> None:
-        if isinstance(self.data, Quantity):
-            self.data = self.data.to(self.units).value
-
-
-@dataclass
-class Transition:
-    data: Data[Literal["transition"], Any]
-    long_name: Attr[str] = "Transition"
-
-
-@dataclass
-class KineticTemperature(Units):
-    data: Data[Literal["T_kin"], float]
-    long_name: Attr[str] = "Kinetic temperature"
-    units: Attr[str] = "K"
-
-
-@dataclass
-class H2Density(Units):
-    data: Data[Literal["n_H2"], float]
-    long_name: Attr[str] = "H2 density"
-    units: Attr[str] = "cm^-3"
-
-
-@dataclass
-class ParaH2Density(Units):
-    data: Data[Literal["n_pH2"], float]
-    long_name: Attr[str] = "Para-H2 density"
-    units: Attr[str] = "cm^-3"
-
-
-@dataclass
-class OrthoH2Density(Units):
-    data: Data[Literal["n_oH2"], float]
-    long_name: Attr[str] = "Ortho-H2 density"
-    units: Attr[str] = "cm^-3"
-
-
-@dataclass
-class ElectronDensity(Units):
-    data: Data[Literal["n_e"], float]
-    long_name: Attr[str] = "Electron density"
-    units: Attr[str] = "cm^-3"
-
-
-@dataclass
-class HydrogenDensity(Units):
-    data: Data[Literal["n_H"], float]
-    long_name: Attr[str] = "Hydrogen density"
-    units: Attr[str] = "cm^-3"
-
-
-@dataclass
-class HeliumDensity(Units):
-    data: Data[Literal["n_He"], float]
-    long_name: Attr[str] = "Helium density"
-    units: Attr[str] = "cm^-3"
-
-
-@dataclass
-class ProtonDensity(Units):
-    data: Data[Literal["n_Hp"], float]
-    long_name: Attr[str] = "Proton density"
-    units: Attr[str] = "cm^-3"
-
-
-@dataclass
-class BackgroundTemperature(Units):
-    data: Data[Literal["T_bg"], float]
-    long_name: Attr[str] = "Background temperature"
-    units: Attr[str] = "K"
-
-
-@dataclass
-class ColumnDensity(Units):
-    data: Data[Literal["N"], float]
-    long_name: Attr[str] = "Column density"
-    units: Attr[str] = "cm^-2"
-
-
-@dataclass
-class LineWidth(Units):
-    data: Data[Literal["dv"], float]
-    long_name: Attr[str] = "Line width"
-    units: Attr[str] = "km s^-1"
-
-
-@dataclass
-class RadexBinary:
-    data: Data[Literal["radex"], str]
-    long_name: Attr[str] = "RADEX binary"
-
-
-@dataclass
-class Line:
-    data: Data[VarDims, str]
-    long_name: Attr[str] = "Line name"
-
-
-@dataclass
-class UpperStateEnergy(Units):
-    data: Data[VarDims, float]
-    long_name: Attr[str] = "Upper state energy"
-    units: Attr[str] = "K"
-
-
-@dataclass
-class Frequency(Units):
-    data: Data[VarDims, float]
-    long_name: Attr[str] = "Frequency"
-    units: Attr[str] = "GHz"
-
-
-@dataclass
-class Wavelength(Units):
-    data: Data[VarDims, float]
-    long_name: Attr[str] = "Wavelength"
-    units: Attr[str] = "um"
-
-
-@dataclass
-class ExcitationTemperature(Units):
-    data: Data[VarDims, float]
-    long_name: Attr[str] = "Excitation temperature"
-    units: Attr[str] = "K"
-
-
-@dataclass
-class OpticalDepth(Units):
-    data: Data[VarDims, float]
-    long_name: Attr[str] = "Optical depth"
-    units: Attr[str] = "dimensionless"
-
-
-@dataclass
-class PeakIntensity(Units):
-    data: Data[VarDims, float]
-    long_name: Attr[str] = "Peak intensity"
-    units: Attr[str] = "K"
-
-
-@dataclass
-class UpperStatePopulation(Units):
-    data: Data[VarDims, float]
-    long_name: Attr[str] = "Upper state population"
-    units: Attr[str] = "dimensionless"
-
-
-@dataclass
-class LowerStatePopulation(Units):
-    data: Data[VarDims, float]
-    long_name: Attr[str] = "Lower state population"
-    units: Attr[str] = "dimensionless"
-
-
-@dataclass
-class IntegratedIntensity(Units):
-    data: Data[VarDims, float]
-    long_name: Attr[str] = "Integrated intensity"
-    units: Attr[str] = "K km s^-1"
-
-
-@dataclass
-class Flux(Units):
-    data: Data[VarDims, float]
-    long_name: Attr[str] = "Flux"
-    units: Attr[str] = "erg s^-1 cm^-2"
-
-
-@dataclass
-class EmptySet(AsDataset):
-    """Specification of an empty dataset."""
-
-    # attributes
-    datafile: Attr[StrPath]
-
-    # dimensions
-    transition: Coordof[Transition]
-    T_kin: Coordof[KineticTemperature]
-    n_H2: Coordof[H2Density]
-    n_pH2: Coordof[ParaH2Density]
-    n_oH2: Coordof[OrthoH2Density]
-    n_e: Coordof[ElectronDensity]
-    n_H: Coordof[HydrogenDensity]
-    n_He: Coordof[HeliumDensity]
-    n_Hp: Coordof[ProtonDensity]
-    T_bg: Coordof[BackgroundTemperature]
-    N: Coordof[ColumnDensity]
-    dv: Coordof[LineWidth]
-    radex: Coordof[RadexBinary]
-
-    # data variables
-    line: Dataof[Line] = field(init=False)
-    E_up: Dataof[UpperStateEnergy] = field(init=False)
-    freq: Dataof[Frequency] = field(init=False)
-    wavel: Dataof[Wavelength] = field(init=False)
-    T_ex: Dataof[ExcitationTemperature] = field(init=False)
-    tau: Dataof[OpticalDepth] = field(init=False)
-    T_R: Dataof[PeakIntensity] = field(init=False)
-    pop_up: Dataof[UpperStatePopulation] = field(init=False)
-    pop_low: Dataof[LowerStatePopulation] = field(init=False)
-    I: Dataof[IntegratedIntensity] = field(init=False)
-    F: Dataof[Flux] = field(init=False)
-
-    def __post_init__(self) -> None:
-        """Set empty arrays to data variables."""
-        model = DataModel.from_dataclass(self)
-        shape = []
-
-        for entry in model.coords:
-            shape.append(len(np.atleast_1d(entry.value)))
-
-        for entry in model.data_vars:
-            setattr(self, str(entry.name), np.empty(shape))
